Move lib/mastodon/premailer_webpack_strategy to lib/
