+++ config/initializers/premailer_rails.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative '../../lib/mastodon/premailer_webpack_strategy'
+require_relative '../../lib/premailer_webpack_strategy'
 
 Premailer::Rails.config.merge!(remove_ids: true,
                                adapter: :nokogiri,
